requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
apscheduler==3.10.4
sendgrid==6.11.0
python-dotenv==1.0.0
